requirements.txt
Django==1.9
django-crispy-forms==1.6.0